+++ backend/src/main/java/com/saas_learning/backend/dto/TransactionResponse.java
@@ -2,10 +2,12 @@ package com.saas_learning.backend.dto;
 
 import com.saas_learning.backend.entities.Transaction;
 
+import java.math.BigDecimal;
+
 public record TransactionResponse(
         String date,
         String description,
-        double montant,
+        BigDecimal montant,
         String categorie
 ) {
     public static TransactionResponse from(Transaction t) {
